requirements.txt
flask
flask-socketio
langchain-groq
python-socketio
websocket-client
streamlit
